Add enumerations, adjust readme and up version (#3)
* Add edge-light export and functionality

Introduced a new "edge-light" export in package.json to point to a newly created edge-light index in the source directory. This includes implementations for parsing, cloning, and reading HTTP message bodies, supporting different media types and handling accordingly.

* Update path for "edge-light" export in package.json

Adjusted the "edge-light" export path in the package.json to reflect its new location within the "edge-light" directory. This ensures the correct file is referenced when the "edge-light" export is used.

* Refactor HTTP message handling in edge-light

Removed unnecessary type checking using the `is` function and streamlined error messages for better clarity and simplicity. Eliminated redundant console logging, focusing on ensuring that HTTP messages have the necessary methods.

* Add function to extract media type from HTTP message

Introduce a `media` function to determine the media type from an HTTP message's headers, ensuring the 'headers' property is present and throwing a TypeError if not. The function retrieves the 'Content-Type' value from the headers or returns an empty string if undefined.

* Remove unnecessary console log in index.js

The console log for 'EdgeRuntime test' has been removed from the read function in index.js. This helps clean up the code by eliminating extraneous logging messages that are not essential for the function's operation.

* Adjust formatting and up version

* Add prototypes and enumerations, adjust readme

* Adjust types and up version

* Move ResponseStatus to it's own package

---------

Co-authored-by: Vladislav Ponomarev <PonomareVlad@users.noreply.github.com>